A small bug on the notation of the color masks has been fixed

+++ objects/graphic/particle.obj.h
@@ -53,7 +53,7 @@ d_declare_class(particle) {
 	struct s_particle_configuration configuration;
 	struct s_particle_information particles[d_particle_cores];
 	t_boolean initialized;
-	unsigned int mask_R, mask_G, mask_B, mask_A;
+	double mask_R, mask_G, mask_B, mask_A;
 } d_declare_class_tail(particle);
 struct s_particle_attributes *p_particle_alloc(struct s_object *self);
 extern struct s_object *f_particle_new(struct s_object *self, struct s_object *drawable_particle, struct s_object *environment,
